Dev to henry ford integration

+++ app/services/v1/intervention/publish.rb
@@ -9,7 +9,6 @@ class V1::Intervention::Publish
   def execute
     timestamp_published_at
     calculate_days_after_schedule
-    delete_draft_answers
     delete_preview_data
   end
 
@@ -42,20 +41,13 @@ class V1::Intervention::Publish
     end
   end
 
-  def delete_draft_answers
-    Answer.joins(question: { question_group: :session }).where(
-      questions: { question_groups: { sessions: { intervention_id: intervention.id } } }
-    ).destroy_all
-  end
-
   def delete_preview_data
     session_ids = intervention.sessions.select(:id)
     preview_users = User.where(preview_session_id: session_ids)
     preview_user_ids = preview_users.select(:id)
-    UserSession.where(user_id: preview_user_ids).delete_all
-    UserLogRequest.where(user_id: preview_user_ids).delete_all
-    Phone.where(user_id: preview_user_ids).delete_all
-    preview_users.delete_all
+    UserIntervention.where(user_id: preview_user_ids).destroy_all
+    UserLogRequest.where(user_id: preview_user_ids).destroy_all
+    preview_users.destroy_all
   end
 
   def selected_number_of_days(session)
